back/requirements.txt
Django==5.0.6
psycopg2-binary>=2.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-yasg==1.21.7
